requirements.txt

asgiref==3.3.1
automium==0.2.6
automium-web==0.1.1
click==7.1.2
codegen==1.0
Django==3.1.7
django-ckeditor==6.0.0
django-ckeditor-staticfiles==4.3.3
django-datetime-widget2==0.9.5
django-js-asset==1.2.2
django-phonenumber-field==5.1.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
legacy==0.1.6
MarkupSafe==1.1.1
patterns==0.3
Pillow==8.2.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
views==0.3
Werkzeug==1.0.1
whitenoise==5.2.0
###
phonenumbers
